Start of Version 2
This version is intended to replace the original and allow extra
functionallity including export to other formats than GoogleDocs (like
and fdf to go with Paizo's fillable character sheet, or DragonberryPi
pawns).
Regardless it already is much more object orientend and already creates
an object tree containing the portfolio of characters with content from
both the XML and the HTML to work around issues of content which is only
present in one or the other.