empty image bug fix, and add test option for flags
Now with the GoogleSlide renderer, if an image match returns nothing, an
empty transparent image is created and inserted.
The ascii "record separator" character: hex 1F can be used to separate
multiple features as part of the value for a match keyword.  Now also
the "group separator": hed 1E can be used to separate a simple
conditional from the feature to insert if the conditional evaluates
true.