added intercalated lists
When a match value string refers to a list and multiple items separated
by \1xe are included the values will be intercalated and joined with a
space.  If the list flag is in the keyword these values will then be
further joined into a list.  Conditionals will operate similarly.  This
way you can get a list of names and values rather then a list of names
followed by a list of values.