Update HeroLabStatRenderGoogleSlide.py
Updated to no longer use access_token as part of url, since Google API does not allow it.
Had to make images public while doing replaceAllShapesWithImage.
Also made it so images are uploaded along side the final Google Presentation, so that when HLRender crashes the scrap files are not left in the top level drice